// src/app/dto/sign-up.dto.ts
export class signUpDto {

    user_firstname: string;

    user_lastname: string;

    user_phone: number;

    user_address: string;

    email: string;

    user_gender: string = "M";

    job_role: number;

    user_password: string;

    // token: string = 'EAUWP13joU';

    token: string | any;
    
}